conditions to acces to modify post

+++ Projet4/alaska/view/frontend/listPostsView.php
@@ -1,6 +1,8 @@
-<?php $title = 'Mon blog'; ?>
-
-<?php ob_start(); ?>
+<?php $title = 'Mon blog';
+if (isset($_SESSION['admin'])) {
+  $admin = $_SESSION['admin'];
+}
+ob_start(); ?>
 <h1>Mon super blog !</h1>
 <p class="navigator">Derniers billets du blog :</p>
 <a href="?action=home">Home</a>
@@ -13,8 +15,18 @@ while ($data = $posts->fetch())
         <h3><a href="index.php?action=post&amp;id=<?= $data['id'] ?>">
             <?= htmlspecialchars($data['title']) ?>
             <em>le <?= $data['creation_date_fr'] ?></em></a>
-            <em><a href="index.php?action=editPost&amp;id=<?= $data['id'] ?>">Modifier post</a></em>
-            <em><a href="index.php?action=deletePost&amp;id=<?= $data['id'] ?>">Supprimer post</a></em>
+            <?php
+            if (isset($admin))
+            {
+              if ($admin == 1)
+              {
+                ?>
+                <em><a href="index.php?action=editPost&amp;id=<?= $data['id'] ?>">Modifier post</a></em>
+                <em><a href="index.php?action=deletePost&amp;id=<?= $data['id'] ?>">Supprimer post</a></em>
+                <?php
+              }
+            }
+            ?>
         </h3>
 
         <p>
